Implement comment CRUD and improve phone details UX
Added create and delete comment functionality to the frontend and backend, including form validation and error handling. Refactored details board and edit components to use observable-based data flow and improved edit mode handling. Updated comment and phone services for better state management. Enhanced UI/UX for comment forms, error messages, and phone details, and fixed several minor bugs in forms and templates. Backend comment controller now returns updated comment lists after create/delete, and enforces minimum comment length.

+++ Rest-api/controllers/commentController.js
@@ -1,65 +1,82 @@
-const mongoose = require('mongoose');
-const { userModel,phoneModel, commentModel } = require('../models');
+const mongoose = require("mongoose");
+const { userModel, phoneModel, commentModel } = require("../models");
 
 function newComment(text, userId, phoneId) {
-    return commentModel.create({ text, userId, phoneId })
-        .then(comment => {
+    return commentModel
+        .create({ text, userId, phoneId })
+        .then((comment) => {
             return Promise.all([
-                userModel.updateOne({ _id: userId }, { $push: { comments: comment._id }, $addToSet: { phones: phoneId } }),
-                phoneModel.findByIdAndUpdate({ _id: phoneId }, { $push: { comments: comment._id }, $addToSet: { buyers: userId } }, { new: true })
-            ])
+                userModel.findById(userId),
+                phoneModel.findById(phoneId),
+            ]).then(([user, phone]) => {
+                if (!user) throw new Error("User not found");
+                if (!phone) throw new Error("Phone not found");
+                return Promise.all([
+                    userModel.updateOne(
+                        { _id: userId },
+                        { $push: { comments: comment._id } }
+                    ),
+                    phoneModel.updateOne(
+                        { _id: phoneId },
+                        { $push: { comments: comment._id } }
+                    ),
+                ]).then(() => {
+                    return commentModel
+                        .find({ phoneId })
+                        .populate("userId", "username")
+                        .sort({ createdAt: -1 });
+                });
+            });
         })
+        .catch((error) => {
+            console.error("Error in newComment:", error);
+            throw error;
+        });
 }
 
 function getLatestsComments(req, res, next) {
     const limit = Number(req.query.limit) || 0;
 
-    commentModel.find()
+    commentModel
+        .find()
         .sort({ created_at: -1 })
         .limit(limit)
-        .populate('userId', 'username')
-        .then(comments => {
-            res.status(200).json(comments)
-          
+        .populate("userId", "username")
+        .then((comments) => {
+            res.status(200).json(comments);
         })
         .catch(next);
 }
-function getCommentsByPhoneId(req,res,next){
+function getCommentsByPhoneId(req, res, next) {
+    const { phoneId } = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(phoneId)) {
+        return res.status(400).json({
+            success: false,
+            message: "Invalid phone ID format",
+        });
+    }
 
-
-  const { phoneId } = req.params;
-    console.log(phoneId);
-    
-  if (!mongoose.Types.ObjectId.isValid(phoneId)) {
-    return res.status(400).json({
-      success: false,
-      message: 'Invalid phone ID format'
-    });
-  }
-
-  commentModel.find({ phoneId })
-    .populate('userId', 'username avatar')
-    .populate('likes', 'username')
-    .sort({ created_at: -1 })
-    .then(comments => {
-  
-      
-      res.status(200).json({
-        success: true,
-        count: comments.length,
-        data: comments
-      });
-    })
-    .catch(error => {
-      console.error('Error:', error);
-      res.status(500).json({
-        success: false,
-        message: 'Server error',
-        error: error.message
-      });
-    });
-
+    commentModel
+        .find({ phoneId })
+        .populate("userId", "username avatar")
+        .populate("likes", "username")
+        .sort({ created_at: -1 })
+        .then((comments) => {
+            res.status(200).json({
+                success: true,
+                count: comments.length,
+                data: comments,
+            });
+        })
+        .catch((error) => {
+            console.error("Error:", error);
+            res.status(500).json({
+                success: false,
+                message: "Server error",
+                error: error.message,
+            });
+        });
 }
 
 function createComment(req, res, next) {
@@ -67,9 +84,9 @@ function createComment(req, res, next) {
     const { _id: userId } = req.user;
     const { commentText } = req.body;
 
-    newComment(commentText, userId, phoneId)
-        .then(([_, updatedPhone]) => res.status(200).json(updatedPhone))
-        .catch(next);
+    newComment(commentText, userId, phoneId).then((updatedComments) =>
+        res.status(200).json(updatedComments)
+    );
 }
 
 function editComment(req, res, next) {
@@ -78,12 +95,16 @@ function editComment(req, res, next) {
     const { _id: userId } = req.user;
 
     // if the userId is not the same as this one of the comment, the comment will not be updated
-    commentModel.findOneAndUpdate({ _id: commentId, userId }, { text: commentText }, { new: true })
-        .then(updatedComment => {
+    commentModel
+        .findOneAndUpdate(
+            { _id: commentId, userId },
+            { text: commentText },
+            { new: true }
+        )
+        .then((updatedComment) => {
             if (updatedComment) {
                 res.status(200).json(updatedComment);
-            }
-            else {
+            } else {
                 res.status(401).json({ message: `Not allowed!` });
             }
         })
@@ -94,17 +115,39 @@ function deleteComment(req, res, next) {
     const { commentId, phoneId } = req.params;
     const { _id: userId } = req.user;
 
-    Promise.all([
-        commentModel.findOneAndDelete({ _id: commentId, userId }),
-        userModel.findOneAndUpdate({ _id: userId }, { $pull: { comments: commentId } }),
-        phoneModel.findOneAndUpdate({ _id: phoneId }, { $pull: { comments: commentId } }),
-    ])
-        .then(([deletedOne, _, __]) => {
-            if (deletedOne) {
-                res.status(200).json(deletedOne)
-            } else {
-                res.status(401).json({ message: `Not allowed!` });
+    commentModel
+        .findOne({ _id: commentId, userId })
+        .then((comment) => {
+            if (!comment) {
+                return res
+                    .status(401)
+                    .json({ message: "Not allowed or comment not found!" });
             }
+
+            return Promise.all([
+                commentModel.findOneAndDelete({ _id: commentId, userId }),
+                userModel.findOneAndUpdate(
+                    { _id: userId },
+                    { $pull: { comments: commentId } }
+                ),
+                phoneModel.findOneAndUpdate(
+                    { _id: phoneId },
+                    { $pull: { comments: commentId } }
+                ),
+            ]);
+        })
+        .then(([deletedComment, _, __]) => {
+            if (!deletedComment) {
+                return res.status(404).json({ message: "Comment not found" });
+            }
+
+            return commentModel
+                .find({ phoneId })
+                .populate("userId", "username")
+                .sort({ createdAt: -1 });
+        })
+        .then((updatedComments) => {
+            res.status(200).json(updatedComments);
         })
         .catch(next);
 }
@@ -113,30 +156,31 @@ function like(req, res, next) {
     const { commentId } = req.params;
     const { _id: userId } = req.user;
 
-    commentModel.findOneAndUpdate(
-        { _id: commentId },
-        { $addToSet: { likes: userId } },
-        { new: true }
-    )
-    .populate('likes', 'username _id') 
-    .populate('userId', 'username _id') 
-    .then(updatedComment => {
-        if (!updatedComment) {
-            return res.status(404).json({ message: 'Comment not found' });
-        }
-        res.status(200).json(updatedComment);
-    })
-    .catch(error => {
-        next(error);
-    });
+    commentModel
+        .findOneAndUpdate(
+            { _id: commentId },
+            { $addToSet: { likes: userId } },
+            { new: true }
+        )
+        .populate("likes", "username _id")
+        .populate("userId", "username _id")
+        .then((updatedComment) => {
+            if (!updatedComment) {
+                return res.status(404).json({ message: "Comment not found" });
+            }
+            res.status(200).json(updatedComment);
+        })
+        .catch((error) => {
+            next(error);
+        });
 }
 
 module.exports = {
-  getLatestsComments,
+    getLatestsComments,
     newComment,
-   createComment,
+    createComment,
     editComment,
     deleteComment,
     like,
     getCommentsByPhoneId,
-}
+};
